Simplify converters structure
