+++ winter/converters/__init__.py
@@ -1,3 +1,3 @@
-from .convert_exception import ConvertException
+from .converter import ConvertException
 from .converter import convert
 from .converter import converter
